인테리어 좌표 float로 변경

+++ Assets/Scripts/GameManager/DBManager.cs
@@ -108,7 +108,7 @@ public class DBManager
         conn.Update(user);
     }
 
-    
+
     public void Update_DesginShopLevel(int level)
     {
         User user = conn.Find<User>(userName);
@@ -348,7 +348,7 @@ public class DBManager
             .Any(x => x.tableID == tableID);
     }
 
-    public bool Insert_Worker(string workerName, int x, int y)
+    public bool Insert_Worker(string workerName, float x, float y)
     {
         try
         {
@@ -358,7 +358,7 @@ public class DBManager
 
             Debug.Log("ID" + newWorkRoom.workerID);
 
-            int affectedRows = conn.Execute( "UPDATE Interior SET ID = ? " +
+            int affectedRows = conn.Execute("UPDATE Interior SET ID = ? " +
                 "WHERE interiorName = ? AND isSet = 1 AND x = ? AND y = ?",
                 newWorkRoom.workerID, workerName, x, y);
 
@@ -387,7 +387,7 @@ public class DBManager
         }
     }
 
-    public bool Set_InteriorItem(string interiorName, int x, int y) // 없던 인테리어 아이템을 좌표에 위치시키는 메서드
+    public bool Set_InteriorItem(string interiorName, float x, float y) // 없던 인테리어 아이템을 좌표에 위치시키는 메서드
     {
         try
         {
@@ -406,7 +406,7 @@ public class DBManager
         }
     }
 
-    public bool Change_InteriorItem_Pos(int beforeX, int beforeY, int afterX, int afterY) // 인테리어 아이템 위치 변경
+    public bool Change_InteriorItem_Pos(float beforeX, float beforeY, float afterX, float afterY) // 인테리어 아이템 위치 변경
     {
         try
         {
@@ -422,7 +422,7 @@ public class DBManager
         }
     }
 
-    public bool NotSet_InteriorItem(int x, int y) // 좌표에 위치되어 있던 인테리어 아이템 빼는 메서드
+    public bool NotSet_InteriorItem(float x, float y) // 좌표에 위치되어 있던 인테리어 아이템 빼는 메서드
     {
         try
         {
@@ -465,4 +465,14 @@ public class DBManager
         tile.tileName = tileName;
         conn.Update(tile);
     }
+
+    //public List<Interior> Select_WallTile_Inventory()
+    //{
+
+    //}
+
+    //public List<Interior> Select_FloorTile_Inventory()
+    //{
+
+    //}
 }
